requirements.txt
langchain
python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
libmagic
OpenAI
pymupdf
langchain-community
langchain_openai 